seperated tracking info data saving logic to repository
